Phase 1: Basic Recycle
Phase 1 – Basic Functionality				
Write a script called recycle that mimics the rm command. The recycle script should be able to accept the name of a file as a command line argument as rm does, but instead of deleting the file, your script should move it to a recyclebin directory called recyclebin located in your home directory.

1.	The script name is recycle and will be stored in $HOME/project.

2.	The recycle bin will be $HOME/recyclebin. Ensure that your script creates this directory.

3.	The file to be deleted will be a command line argument and the script should be executed as follows: 	bash recycle fileName 

4.	The script must test for the following error conditions. When any of the following are encountered, the script must display the same kind of error messages as the rm command, and the script must set a non-zero exit status to terminate.
a)	No filename provided 
b)	File does not exist 
c)	Directory name provided instead of a filename
d)	Filename provided is recycle. In this case, display the error message “Attempting to delete recycle – operation aborted” and terminate the script with a non-zero exit status. Do not generate this error if the filename is only similar to recycle. For example, a file called recycle2015 should not generate any error messages. Be sure to create a hard link for your script before testing this error.

5.	The filenames in the recyclebin, will be in the following format:
fileName_inode
For example, if a file named f1 with inode 1234 is recycled, the file in the recyclebin will be named f1_1234. This gets around the potential problem of deleting two files with the same name. The recyclebin will only contain files, not directories. 

6.	Create a hidden file called .restore.info in $HOME.  Each line of this file will contain the name of the file in the recyclebin, followed by a colon, followed by the original absolute path of the file.  For example, if a file called f1, with an inode of 1234 was recycled from the /home/trainee1 directory, .restore.info will contain:
f1_1234:/home/trainee1/f1
If  another file named f1, with an inode of 5432, was recycled from the /home/trainee1/testing directory, then .restore.info will contain:
f1_1234:/home/trainee1/f1
f1_5432:/home/trainee1/testing/f1